Write creds files out to windows and *nix variants
