+++ main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/user"
 	"path"
 	"strconv"
 	"strings"
@@ -314,26 +313,6 @@ const defaultDotDir = ".carina"
 
 // CredentialsBaseDirEnvVar environment variable name for where credentials are downloaded to by default
 const CredentialsBaseDirEnvVar = "CARINA_CREDENTIALS_DIR"
-
-const xdgDataHomeEnvVar = "XDG_DATA_HOME"
-
-// CarinaCredentialsBaseDir get the current base directory for carina credentials
-func CarinaCredentialsBaseDir() (string, error) {
-	if os.Getenv(CredentialsBaseDirEnvVar) != "" {
-		return os.Getenv(CredentialsBaseDirEnvVar), nil
-	}
-
-	// Support XDG
-	if os.Getenv(xdgDataHomeEnvVar) != "" {
-		return path.Join(os.Getenv(xdgDataHomeEnvVar), "carina"), nil
-	}
-
-	currentUser, err := user.Current()
-	if err != nil {
-		return "", err
-	}
-	return path.Join(currentUser.HomeDir, defaultDotDir), nil
-}
 
 // Create a cluster
 func (carina *CreateCommand) Create(pc *kingpin.ParseContext) (err error) {
